// src/api/hooks/user/key.ts
export enum UserQueryEnum {
  UserAll = 'use-get-all-users',
  UserGet = 'use-get-user',
  UserSave = 'use-save-user',
  UserUpdate = 'use-update-user',
  UserStatusChange = 'use-status-change',
  UserDelete = 'use-delete-user',
  UserGetRole = 'user-role',
  UserPasswordUpdate = 'user-password-update',
}